# backend/src/main/java/com/rahil/book_nexus/auth/AuthenticationResponse.java
package com.rahil.book_nexus.auth;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthenticationResponse {
    private String token;

}