# backend/src/main/java/com/rahil/book_nexus/book/BorrowedBookResponse.java
package com.rahil.book_nexus.book;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
    private boolean read;
    private int readCount;
    private byte[] cover;
}