# backend/src/main/java/com/rahil/book_nexus/feedback/FeedbackResponse.java
package com.rahil.book_nexus.feedback;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double rating;
    private String review;
    private boolean ownFeedback;
}